pyproject.toml
[tool.poetry]
name = "relia-cli"
version = "1.0.0"
description = "Relia OSS: Local-first, self-hosted Ansible playbook generator"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10"
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
typer = "^0.9.0"
requests = "^2.31.0"
ansible-lint = { version = "^6.22.0", markers = "platform_system != 'Windows'" }
pyjwt = "^2.8.0"
tenacity = "^8.2.3"
openai = "^1.14.0"
boto3 = "^1.34.64"
structlog = "^24.1.0"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
pytest = "^8.0.2"
pytest-asyncio = "^0.23.5"
httpx = "^0.27.0"
psutil = "^5.9.8"
jinja2 = "^3.1.2"

[tool.poetry.scripts]
relia-cli = "relia_cli.main:app"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.6"
pytest-cov = "^4.1.0"
coverage = "^7.4.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["backend", "relia_cli"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/conftest.py",
    "*/wsgi.py",
    "*/asgi.py",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]
precision = 2
fail_under = 75
show_missing = true
skip_covered = true
